Move configuration to sample file
The user of the application may need to configure their `todo.txt` and
`done.txt` paths. Configuration of these paths should not be done in
a checked in file.

This should probably be something that's read from the environment
variables instead, but for now this will do.

Tweak setup instructions in README.

Include .gitignore file, to ensure that `config/config.exs` does not
show up as an untracked file.