Add hovering over panel elements
Allows user to scroll through the elements in a selected panel.

Duplicates the private function `find_and_update/3`; this function
should probably be extracted into some helper module.

The private functions starting with `shift_hover` are essentially copies
of the similar private functions found in the `Update.Normal` module.
This kind of shifting of values seems to be a repeated concept. Could be
worth extracting to a helper module.

Chosen colors are temporary, just used something to make the hovering
clear.