Fix panel hover index when adding/removing panels
The `model.todo_panel_hover_index` is currently not adjusted upon the
events of adding/removing panels, leading to surprising situations, e.g.
where the recently added panel looks like it's hovered over, but
pressing Enter results in the panel under the unchanged hover index
position to be selected.

The current working of the application may also lead to out-of-bounds
positioning of the hover index.

This patch introduces the assumption that adding/removing a panel leads
to moving the hover index to the top.