Change how panel element hover works
Currently the shifting of which `TodoPanel.Element` has the `hover`
value set to `true` is done with a recursive function that pushes the
hover value around. This is a remnant from before hover indices were
introduced to `%Model{}` and `%TodoPanel{}`.

This patch introduces the following changes:

  1. Add the `element_hover_index` field to `TodoPanel` structs.
  2. Replace the recursive shifting of the hover value, to instead set
     all the elements to not be hovered, then set the one indicated by
     the index to be hovered.
  3. Simplify the helper function `find_and_update/3` to
     `update_selected_panel/2`, as finding the selected panel was the
     only kind of finding done.

Ironically, changing the logic to rely on the index instead of
traversing the panel elements, has resulted in the amount of lines added
for the index shifting to be similar to the amount of lines removed for
the hover value shifting.