Use Taskwarrior hex for parsing tasks
Read tasks from the command `task export` and store the results as
`TodoBoard.Todo` structs. Since the `TodoBoard.Todo` struct was so far
unused, change its fields and bring it closer to what TaskWarrior todos
offer in terms of fields.

Introduce the `Repo` interface, for a generic way of reading and writing
`TodoBoard.Todo` items. My intention for the foreseeable future is to
work with the TaskWarrior utility, but in case someone else would want
to use a different backend, the `Repo` interface should allow for
replacing TaskWarrior in a rather straightforward fashion.

Currently the code for writing TODO changes to TaskWarrior is not
implemented (only stub code). This is not a problem at the moment, due
to there being no functionality for mutating data. Will need to be
addressed in the future.

Error handling for reading the TODOs from TaskWarrior is non-existent.
Some kind of handling should be added (at the very least to conform to
the callback type spec).