Extract rendering of listing into Listing context
To get closer to the goal of having separate tabs with separate content,
move the parts of the `App.render/1` function into the `Listing`
context, specifically to the `App.Listing.Render` module.

This way it's easier to read the `App.render/1` function and see what
code pertains to rendering the `Listing` related view elements, and what
is there for the surrounding application.